decode signed timestamps correctly for AFP 2.x sessions
AFP date fields are signed seconds relative to 2000-01-01. Decode them before applying the epoch and retain the AFP 2.x server-to-client clock offset from FPGetSrvrParms. Use the unset-date sentinel only as a directory fallback, preserving real timestamps.

Widen public and IPC date fields to retain valid pre-2000 values and add protocol coverage for signed values, the offset, and the sentinel.

+++ daemon/stateless_ipc.h
@@ -499,8 +499,8 @@ struct afpsl_ipc_desktop_appl_response {
     uint32_t tag;
     uint32_t directory_id;
     uint32_t file_id;
-    uint32_t creation_date;
-    uint32_t modification_date;
+    int64_t creation_date;
+    int64_t modification_date;
     uint64_t data_fork_size;
     uint64_t resource_fork_size;
     char pathname[AFPC_MAX_NAME_BYTES];
